+++ src/hal/helpers/overhead.py
@@ -134,9 +134,13 @@ class OverheadHelper:
             )
             return
 
-        current_macro_id = Overhead.select(peewee.fn.MAX(Overhead.macro_id)).scalar()
+        try:
+            macro_id_last = Overhead.select(peewee.fn.MAX(Overhead.macro_id)).scalar()
+        except Exception as err:
+            warnings.warn(f"Failed getting macro_id: {err}", HALWarning)
+            return
 
-        return (current_macro_id or 0) + 1
+        return (macro_id_last or 0) + 1
 
     def update_database(self):
         """Updates the database with the overhead."""