+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+
+## Next version
+
+### 🚀 New
+
+* [#17](https://github.com/sdss/HAL/pull/17) Use 730 second exposure times for bright designs.
+
+
 ## 1.0.3 - March 5, 2024
 
 ### 🏷️ Changed